+++ packages/ethereum-indexer-db-processors/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ethereum-indexer-db-processors
 
+## 0.6.7
+
+### Patch Changes
+
+- improve processor import to work in pnpm + startBlock fix
+- Updated dependencies
+  - ethereum-indexer-db-utils@0.6.7
+  - ethereum-indexer@0.6.7
+
 ## 0.6.6
 
 ### Patch Changes